requirements.txt
asgiref==3.3.4
boto3==1.18.24
botocore==1.21.24
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
decorator==5.0.9
Django==3.2.4
django-cors-headers==3.7.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
dynaconf==3.1.4
future==0.18.2
hvac==0.11.0
idna==3.2
Jinja2==3.0.1
jmespath==0.10.0
knox==0.1.14
loguru==0.5.3
MarkupSafe==2.0.1
numpy
pandas
probableparsing==0.0.1
psycopg2
pyap==0.3.1
pycparser==2.20
Pygments==2.9.0
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
s3transfer==0.5.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
validators==0.18.2
